+++ client/src/components/Header/index.js
@@ -1,0 +1,2 @@
+import component from "./Header";
+export default component;
